requirements.txt
fastapi
uvicorn
ollama
python-dotenv
requests
pandasai
openai
bamboo